Close idle connection with no inbound traffic after certain time. Configurable idle time HTTP_CONNECTION_READ_IDLE_TIME_SECONDS
