--- blueflood-http/src/main/java/com/rackspacecloud/blueflood/inputs/handlers/IdleStateEventHandler.java
@@ -1,25 +1,0 @@
-package com.rackspacecloud.blueflood.inputs.handlers;
-
-import io.netty.channel.ChannelDuplexHandler;
-import io.netty.channel.ChannelHandlerContext;
-import io.netty.handler.timeout.IdleState;
-import io.netty.handler.timeout.IdleStateEvent;
-import org.slf4j.Logger;
-import org.slf4j.LoggerFactory;
-
-
-public class IdleStateEventHandler extends ChannelDuplexHandler {
-
-    private static final Logger log = LoggerFactory.getLogger(IdleStateEventHandler.class);
-
-    @Override
-    public void userEventTriggered(ChannelHandlerContext ctx, Object evt) throws Exception {
-        if (evt instanceof IdleStateEvent) {
-            IdleStateEvent e = (IdleStateEvent) evt;
-            if (e.state() == IdleState.READER_IDLE) {
-                log.info("Connection is closed as there is no inbound traffic, Connection: [" + ctx.channel().toString() + "]");
-                ctx.close();
-            }
-        }
-    }
-}
